+++ src/main/java/yesman/epicfight/api/client/neoevent/MappedMovementInputUpdateEvent.java
@@ -7,7 +7,7 @@ import org.jetbrains.annotations.Nullable;
 import net.minecraft.client.player.Input;
 import yesman.epicfight.api.client.input.PlayerInputState;
 import yesman.epicfight.api.client.input.handlers.InputManager;
-import yesman.epicfight.api.client.input.utils.InputUtils;
+import yesman.epicfight.client.input.InputUtils;
 import yesman.epicfight.api.neoevent.playerpatch.PlayerPatchEvent;
 import yesman.epicfight.client.world.capabilites.entitypatch.player.LocalPlayerPatch;
