+++ src/main/java/yesman/epicfight/skill/dodge/DodgeSkill.java
@@ -1,6 +1,5 @@
 package yesman.epicfight.skill.dodge;
 
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.util.Mth;
@@ -14,7 +13,7 @@ import yesman.epicfight.api.animation.types.StaticAnimation;
 import yesman.epicfight.api.client.camera.EpicFightCameraAPI;
 import yesman.epicfight.api.client.input.MovementDirection;
 import yesman.epicfight.api.client.input.handlers.InputManager;
-import yesman.epicfight.api.client.input.utils.InputUtils;
+import yesman.epicfight.client.input.InputUtils;
 import yesman.epicfight.client.events.engine.ControlEngine;
 import yesman.epicfight.client.world.capabilites.entitypatch.player.LocalPlayerPatch;
 import yesman.epicfight.skill.Skill;
